refactor(control): deplacer Role dans le control engine
Casse la dependance bidirectionnelle control <-> server/rpc.
Role est un concept RBAC du control plane, pas du transport RPC.
Le re-export dans server::rpc::types preserve la compatibilite interne.
